replicaCodice/test_SAVIA/LLM/requirements.txt
accelerate==1.0.1
anaconda-anon-usage==0.4.4
archspec==0.2.3
asttokens==2.0.5
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
bitsandbytes==0.44.1
blinker==1.8.2
boltons==23.0.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
conda==23.5.2
conda-build==24.3.0
conda-content-trust==0.2.0
conda_index==0.4.0
conda-libmamba-solver==23.7.0
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==42.0.5
decorator==5.1.1
distro==1.8.0
dnspython==2.6.1
exceptiongroup==1.2.0
executing==0.8.3
expecttest==0.2.1
filelock==3.13.1
Flask==3.0.3
fsspec==2024.3.1
gmpy2==2.1.2
huggingface-hub==0.26.1
hypothesis==6.100.1
idna==3.4
ipython==8.20.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lark==1.1.9
libarchive-c==2.9
libmambapy==1.5.8
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
menuinst==2.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.4
optree==0.11.0
packaging==23.2
parso==0.8.3
peft==0.13.2
pexpect==4.8.0
pillow==10.2.0
pip==23.3.1
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.0.0
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.6
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==24.0.0
PySocks==1.7.1
python-etcd==0.4.5
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2024.9.11
requests==2.31.0
rpds-py==0.10.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.4.5
setuptools==68.2.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
tokenizers==0.20.1
tomli==2.0.1
toolz==0.12.0
torch==2.3.0
torchaudio==2.3.0
torchelastic==0.2.2
torchvision==0.18.0
tqdm==4.65.0
traitlets==5.7.1
transformers==4.45.2
triton==2.3.0
truststore==0.8.0
types-dataclasses==0.6.6
typing_extensions==4.11.0
urllib3==2.1.0
wcwidth==0.2.5
Werkzeug==3.0.4
wheel==0.41.2
zstandard==0.19.0
